Fix to failing 'replace file' upload option.
Replacing files was failing due to an assumption on uninitialized
file data in the clear cache function. Now it is checked if file
data has been initialized on the clear cache function.

+++ ci_2.0.3_application/libraries/File_lib.php
@@ -142,7 +142,7 @@ class File_lib {
 			while (false !== ($dirfile = readdir($handle))) 
 			{
 				//if there is a file that has additional naming after filename.fileextension, delete it
-				if (strpos($dirfile, $this->file['fullfilename'].'.') === 0) 
+				if (isset($this->file['fullfilename']) && strpos($dirfile, $this->file['fullfilename'].'.') === 0) 
 				{
 					unlink($this->folder.$dirfile);
 				}
